# backend/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Service Tracking Platform API",
        "description": "REST API for the Service Tracking Platform with AI powered service request classification and technician assignment.\n        \n        Features:\n        - JWT-based authentication and authorization\n        - AI Smart Request Creation\n        - AI Technician Recommendation\n        - Status Transition Management\n        - Real-time Assignment Workflow\n        ",
        "contact": {
            "name": "Mileta Lvova",
            "email": "mileta279@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/technicians/assigned-requests": {
            "get": {
                "tags": ["Technicians"],
                "summary": "Get all assigned requests",
                "description": "Endpoint to get all assigned requests",
                "responses": {
                    "200": {
                        "description": "List of assigned requests retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of assigned requests"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Technician"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden",
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/technicians/{id}/status": {
            "patch": {
                "tags": ["Technicians"],
                "summary": "Updates a status of the assignment",
                "description": "Endpoint to update a status of the assignment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "integer"
                                },
                                "example": {
                                    "type": "number",
                                    "example": 1
                                }
                            }
                        },
                        "description": "Service ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Status updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Technician"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden",
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound",
                        "description": "Not Found"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatusUpdateInput"
                            }
                        }
                    }
                }
            }
        },
        "/api/technicians/profile": {
            "patch": {
                "tags": ["Technicians"],
                "summary": "Updates a profile",
                "description": "Endpoint to update a profile",
                "responses": {
                    "200": {
                        "description": "Technician profile successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TechnicianProfile"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden",
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "description": "Not Found",
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TechnicianUpdateInput"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Authorization header using the Bearer scheme. \n                Enter ONLY your JWT token in the input below (without 'Bearer ' prefix).\n                \n                Example: \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\""
            }
        },
        "schemas": {
...
            "Technician": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 1
                                    }
                                }
                            },
                            "assignmentId": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 1
                                    }
                                }
                            },
                            "Assigned": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "2026-05-29T12:30:00.000Z"
                                    }
                                }
                            },
                            "Created": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "2026-05-29T12:30:00.000Z"
                                    }
                                }
                            },
                            "Updated": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "2026-05-29T12:30:00.000Z"
                                    }
                                }
                            },
                            "technicianId": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 1
                                    }
                                }
                            },
                            "FullName": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                }
                            },
                            "Email": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    }
                                }
                            },
                            "Customer": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "Richard Miles"
                                    }
                                }
                            },
                            "CustomerEmail": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "richardmiles@example.com"
                                    }
                                }
                            },
                            "Service": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "Plumbing"
                                    }
                                }
                            },
                            "Address": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "123 Main St, New York, NY, 10001"
                                    }
                                }
                            },
                            "Status": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "Assigned"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "StatusUpdateInput": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "required": {
                        "type": "array",
                        "example": ["statusId"],
                        "items": {
                            "type": "string"
                        }
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "statusId": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 3
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "TechnicianUpdateInput": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "skills": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "Plumbing, Electrical"
                                    }
                                }
                            },
                            "isAvailable": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "boolean"
                                    },
                                    "example": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                }
                            },
                            "currentLocationId": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 1
                                    }
                                }
                            },
                            "maxActiveJobs": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 3
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized - Authentication required",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden - Insufficient permissions",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "JWT": []
        }
    ]
}